fix truncation
git-svn-id: svn+ssh://market.scs.stanford.edu/home/u1/asbestos/repo/jos64/trunk@7441 927ebb28-5305-0410-b3ca-c4b640df69b7

+++ lib/jcomm.cc
@@ -72,7 +72,7 @@ jlink_copyto(struct jlink *jl, const void *buf, uint64_t cnt)
 uint64_t
 jlink_read_bytes(struct jlink *jl)
 {
-    int r;
+    uint64_t r;
     jthread_mutex_lock(&jl->mu);
     r = jl->bytes;
     jthread_mutex_unlock(&jl->mu);
